requirements.txt
streamlit==1.32.0
python-nmap==0.7.1
reportlab==4.1.0
pandas==2.1.0
